Add files via upload
After placing a sample in a regular microscope, that slide needs to be adjusted so that clear image is viewed from the other end. This code is an idea to automate the process of manual adjustment and even to get clear and focused results as output in a simple and easy way. Blurred Images don't have high Laplacian variance, this is taken as a factor and the code is developed to tackle the problem and get clear and focused results with very little manual intervention. In this code with the help of repeated feedback, motor is made to stop in the position where most focused images can be taken and even viewed.