Add files via upload
This Python script controls two stepper motors connected to a Raspberry Pi using the RPi.GPIO library. The script allows the user to specify the number of cycles, the number of steps for each motor, the rotation speed in RPM, and the direction of rotation for each motor. Additionally, the script includes functionality to interrupt the process using the Escape key.